identity: break out cache metrics declarations
